pyproject.toml
[project]
name = "voicetranslator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elevenlabs>=2.0.0",
    "faster-whisper>=1.1.1",
    "gradio>=5.30.0",
    "python-dotenv>=1.1.0",
    "translate>=3.6.1",
]
